chore: update carousel based on PR review
