+++ client/src/pages/HomePage/HomePage.tsx
@@ -1,64 +1,6 @@
 import Carousel from "../../components/Carousel/Carousel";
 import "./HomePage.css";
-
-const photos = [
-  {
-    id: "ph1",
-    src: "surf.webp",
-    caption: "Sports",
-  },
-  {
-    id: "ph2",
-    src: "pexels-christian.webp",
-    caption: "Paysages",
-  },
-  {
-    id: "ph3",
-    src: "pexels-caiquethecreator.webp",
-    caption: "Portraits",
-  },
-];
-
-const dessins = [
-  {
-    id: "d1",
-    src: "claudio-schwarz.webp",
-    caption: "Abstraits",
-  },
-  {
-    id: "d2",
-    src: "illustrated-watercolor.webp",
-    caption: "Villes",
-  },
-  {
-    id: "d3",
-    src: "jakub-durian.webp",
-    caption: "Portraits",
-  },
-];
-
-const peintures = [
-  {
-    id: "p1",
-    src: "240_F2.webp",
-    caption: "Portraits",
-  },
-  {
-    id: "p2",
-    src: "Oeuvre-homePage.webp",
-    caption: "Abstraits",
-  },
-  {
-    id: "p1",
-    src: "240_F2.webp",
-    caption: "Portraits",
-  },
-  {
-    id: "p2",
-    src: "Oeuvre-homePage.webp",
-    caption: "Abstraits",
-  },
-];
+import { photos, dessins, peintures } from "../data/data";
 
 function HomePage() {
   return (
