+++ client/src/pages/data/data.ts
@@ -1,0 +1,58 @@
+export const photos = [
+  {
+    id: "ph1",
+    src: "surf.webp",
+    caption: "Sports",
+  },
+  {
+    id: "ph2",
+    src: "pexels-christian.webp",
+    caption: "Paysages",
+  },
+  {
+    id: "ph3",
+    src: "pexels-caiquethecreator.webp",
+    caption: "Portraits",
+  },
+];
+
+export const dessins = [
+  {
+    id: "d1",
+    src: "claudio-schwarz.webp",
+    caption: "Abstraits",
+  },
+  {
+    id: "d2",
+    src: "illustrated-watercolor.webp",
+    caption: "Villes",
+  },
+  {
+    id: "d3",
+    src: "jakub-durian.webp",
+    caption: "Portraits",
+  },
+];
+
+export const peintures = [
+  {
+    id: "p1",
+    src: "240_F2.webp",
+    caption: "Portraits",
+  },
+  {
+    id: "p2",
+    src: "Oeuvre-homePage.webp",
+    caption: "Abstraits",
+  },
+  {
+    id: "p3",
+    src: "240_F2.webp",
+    caption: "Portraits",
+  },
+  {
+    id: "p4",
+    src: "Oeuvre-homePage.webp",
+    caption: "Abstraits",
+  },
+];
