+++ client/src/types/carousel.ts
@@ -1,0 +1,4 @@
+interface CarouselProps {
+  title: string;
+  items: { id: string; src: string; caption: string }[];
+}
